Make the .atom feed validate.
Switch:

    <author>Drew Neil</author>

to

    <author>
        <name>Drew Neil</name>
    </author>

And change the attribute tag on <summary> from xhtml="true" to
type="text". Use raw textile input, rather than textilizing it to HTML
output. The html_escape command is automatically run on the HTML output,
causing <p> tags to appear in the atom feed. Also, any reference to
radius tags (relatively common in descriptions of Radiant extensions) is
automatically escaped, so that e.g. <r:if> appears as &lt;r:if&gt; in
the atom feed.

[Plain text][1] is suitable for atom feed summary. If the description
makes reference to any HTML or radius tags, they are still escaped, so
appear appropriately in the output.

[1]:
http://www.atomenabled.org/developers/syndication/atom-format-spec.php#rfc.section.3.1.1.1